CLEAN COMMIT TO MAKE A BRANCH FROM

+++ src/server/TcpServer.cpp
@@ -10,9 +10,9 @@ TcpServer::TcpServer(const std::vector<int> & ports) : _ports(ports)
 
     std::cout << G << IT << "Server initialized on port(s): ";
 
-    for (int i = 0; i < static_cast<int>(_ports.size()); ++i) {
+    for (size_t i = 0; i < _ports.size(); ++i) {
         std::cout << _ports[i];
-        if (i < static_cast<int>(_ports.size()) - 1)
+        if (i < _ports.size() - 1)
             std::cout << ", ";
 
     }
@@ -22,7 +22,7 @@ TcpServer::TcpServer(const std::vector<int> & ports) : _ports(ports)
 
 TcpServer::~TcpServer()
 {
-    for (int i = 0; i < static_cast<int>(_serverSockets.size()); ++i) {
+    for (size_t i = 0; i < _serverSockets.size(); ++i) {
         if (_serverSockets[i] != -1)
             close(_serverSockets[i]);
     }
@@ -31,7 +31,7 @@ TcpServer::~TcpServer()
 
 void TcpServer::setupSocket()
 {
-    for (int i = 0; i < static_cast<int>(_ports.size()); ++i) {
+    for (size_t i = 0; i < _ports.size(); ++i) {
         int serverSocket = socket(AF_INET, SOCK_STREAM, 0); /* AF_INET : IPv4, SOCK_STREAM : Socket de connexion, 0 : Protocole TCP */
         if (serverSocket < 0)
         {
@@ -92,12 +92,12 @@ void TcpServer::startServer()
             std::cerr << R << IT << "Poll failed" << RES << std::endl;
             exitCloseFds(_serverSockets);
         }
-        for (int i = 0; i < static_cast<int>(_pollFds.size()); ++i)
+        for (size_t i = 0; i < _pollFds.size(); ++i)
         {
             if (_pollFds[i].revents & POLLIN)
             {
                 bool isServerSocket = false;
-                for (int j = 0; j < static_cast<int>(_serverSockets.size()); ++j)
+                for (size_t j = 0; j < _serverSockets.size(); ++j)
                 {
                     if (_pollFds[i].fd == _serverSockets[   j])
                     {
@@ -163,7 +163,7 @@ void TcpServer::handleClient(int clientFd)
 
 void TcpServer::cleanupClient(int fd)
 {
-    for (int i = 0; i < static_cast<int>(_pollFds.size()); ++i) {
+    for (size_t i = 0; i < _pollFds.size(); ++i) {
         if (_pollFds[i].fd == fd) {
             _pollFds.erase(_pollFds.begin() + i);
             break;
